requirements.txt
# Core ML/NLP Libraries
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
spacy>=3.7.0
language-tool-python>=2.7.1

# Grammar and Text Analysis
nltk>=3.8.1
textstat>=0.7.3
py-readability-metrics>=1.4.5

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
noisereduce>=3.0.0

# API and Web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
pyyaml>=6.0.1
requests>=2.31.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
